app/requirements.txt
mysql-connector-python==8.0.32
openai==0.27.2
python-decouple==3.7
streamlit==1.18.1
